+++ packages/nimble-components/src/utilities/types.ts
@@ -1,0 +1,5 @@
+import type { FoundationElement } from '@microsoft/fast-foundation';
+
+export interface UpdaTable extends FoundationElement {
+    update(): void;
+}
